@@ -11,7 +11,7 @@
         }
     ],
     "require": {
-        "php": ">=5.5.9",
+        "php": ">=7.0",
         "symfony/dependency-injection": ">=2.7 <5.0.0",
         "openlss/lib-array2xml": "~0.0",
         "drupal/core-plugin": "^8.5",
@@ -19,7 +19,7 @@
         "hanneskod/classtools": "~1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.8",
+        "phpunit/phpunit": "~6.5",
         "kzykhys/php-csv-parser": "~1.4"
     },
     "suggest": {
